Random changes
This wasn't commited for a while :-(

Changes:
- Split files into smaller chunks
- Use C++17
- Works on MSVC
