+++ src/CMakeLists.txt
@@ -1,29 +1,21 @@
 project(zusi3tcplib VERSION 0.1 LANGUAGES C CXX)
 
-if (WIN32)
-    #set ( SOURCES Zusi3TCP.cpp WinsockBlockingSocket.cpp )
-    #set ( HEADERS include/Zusi3TCP.h include/WinsockBlockingSocket.h )
-    set( WINDOWS_LIB ws2_32 ) # might only be needed with mingw
-endif()
-#else()
-    set ( SOURCES Zusi3TCP.cpp )
-    set ( HEADERS include/Zusi3TCP.h )
-#endif()
+set ( SOURCES Zusi3TCP.cpp )
+set ( HEADERS include/Zusi3TCP.h include/Zusi3TCPTypes.h include/Zusi3TCPData.h )
 
 
 set ( CLI_SOURCES Zusi3TCPCli.cpp )
 set ( CLI_HEADERS )
 
-set ( SERVER_SOURCES serverconnection.cpp )
-set ( SERVER_HEADERS include/serverconnection.h )
 
-find_package( Boost REQUIRED system )
+find_package( Boost REQUIRED system date_time regex)
 find_package( Threads REQUIRED )
 
 
 add_library(zusi3tcp  ${SOURCES} ${HEADERS} )
 target_include_directories(zusi3tcp PUBLIC ${Boost_INCLUDE_DIRS})
-target_link_libraries(zusi3tcp PUBLIC ${WINDOWS_LIB} ${CMAKE_THREAD_LIBS_INIT} ${Boost_LIBRARIES} ) #Boost::system )
+LINK_DIRECTORIES(${Boost_LIBRARY_DIRS})
+target_link_libraries(zusi3tcp PUBLIC ${CMAKE_THREAD_LIBS_INIT} Boost::system Boost::date_time Boost::regex)
 
 add_executable(zusi3cli ${CLI_SOURCES} ${CLI_HEADES} )
 target_link_libraries(zusi3cli zusi3tcp )
